Details:
Route:

Added a new route in photoRoutes.js:
GET /api/photos/:petId/signed-urls - This endpoint triggers the process to generate signed URLs for all photos associated with a given pet ID.
Controller:

Implemented a new controller function in photoController.js:
generatePhotoUrlsByPetId: This controller function handles the HTTP request to generate signed URLs for all photos related to a specific pet. It retrieves the pet ID from the request parameters, invokes the service function to generate the signed URLs, and returns these URLs in the HTTP response.
Service:

Created a new service function in googleCloudStorage.js:
generateSignedUrls(filenames): This async function takes an array of filenames as input, generates signed URLs for each file stored in Google Cloud Storage, and returns an array of these signed URLs.
Purpose:
The purpose of these changes is to enable the generation and publication of signed URLs for photos stored in Google Cloud Storage. This feature allows users to securely access photos associated with a pet by generating time-limited URLs that can be shared or accessed as needed.